@@ -1,6 +1,6 @@
 {
   "name": "@dxworks/depinder",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "description": "",
   "keywords": [
     "dxworks",
@@ -15,6 +15,9 @@
     "url": "git+https://github.com/dxworks/depinder.git"
   },
   "license": "Apache-2.0",
+  "engines": {
+    "node": ">=24.0.0"
+  },
   "author": "Mario Rivis <mario.rivis@gmail.com>",
   "main": "dist/depinder.js",
   "bin": {
@@ -29,8 +32,8 @@
     "clean:modules": "rm -rf ./node_modules ./package-lock.json",
     "copy-files": "copyfiles -a -u 1 src/assets/**/*.* dist",
     "create": "npm run build && npm run test",
-    "lint": "eslint ./src --ext .ts",
-    "lint:fix": "eslint ./src --ext .ts --fix",
+    "lint": "eslint ./src",
+    "lint:fix": "eslint ./src --fix",
     "local": "npm run build && npm link && npx dx-depinder --help",
     "refresh": "rm -rf ./node_modules ./package-lock.json && npm install",
     "test": "jest --passWithNoTests",
@@ -38,63 +41,60 @@
   },
   "dependencies": {
     "@dxworks/nuget-inspector": "^1.2.0",
-    "@octokit/graphql": "^4.8.0",
+    "@snyk/dep-graph": "^2.13.0",
     "@snyk/gemfile": "^1.2.0",
+    "axios": "^1.7.0",
     "chalk": "^4.1.2",
     "cli-progress": "^3.11.2",
     "commander": "^14.0.2",
-    "cross-env": "^7.0.3",
+    "cross-env": "^10.0.0",
     "csv": "^6.0.4",
-    "dotenv": "^16.0.3",
+    "dotenv": "^17.0.0",
+    "fast-xml-parser": "^5.3.3",
     "lodash": "^4.17.21",
-    "minimatch": "^6.1.6",
-    "moment": "^2.29.1",
-    "mongodb": "^5.8.1",
-    "mongoose": "^7.5.2",
-    "node-html-parser": "^6.1.4",
-    "npm": "^9.9.0",
-    "npm-registry-fetch": "^12.0.1",
-    "packageurl-js": "^0.0.5",
-    "pom-parser": "^1.2.0",
-    "puppeteer": "18.0.1",
-    "semver": "^7.3.5",
-    "snyk-gradle-plugin": "^3.25.2",
-    "snyk-nodejs-lockfile-parser": "^1.60.1",
+    "minimatch": "^10.0.0",
+    "moment": "^2.30.0",
+    "mongodb": "^7.0.0",
+    "mongoose": "^9.0.0",
+    "node-html-parser": "^7.0.0",
+    "npm": "^11.0.0",
+    "npm-registry-fetch": "^19.0.0",
+    "puppeteer": "^24.0.0",
+    "semver": "^7.7.0",
+    "snyk-nodejs-lockfile-parser": "^2.5.0",
     "spdx-correct": "^3.1.1",
     "spdx-license-ids": "^3.0.10",
     "tmp": "^0.2.1",
     "toml": "^3.0.0",
-    "winston": "^3.11.0",
-    "xml2js": "^0.6.2"
+    "winston": "^3.17.0"
   },
   "devDependencies": {
-    "@types/cli-progress": "^3.9.2",
-    "@types/dockerode": "^3.3.0",
-    "@types/fs-extra": "^9.0.13",
-    "@types/inquirer": "^8.1.3",
-    "@types/jest": "^27.0.2",
-    "@types/lodash": "^4.14.177",
-    "@types/node": "^16.10.2",
-    "@types/npm-registry-fetch": "^8.0.3",
-    "@types/semver": "^7.3.9",
-    "@types/spdx-correct": "^2.0.0",
-    "@types/string-template": "^1.0.2",
-    "@types/tmp": "^0.2.3",
-    "@types/ws": "^8.5.4",
-    "@types/xml2js": "^0.4.14",
-    "@typescript-eslint/eslint-plugin": "^4.32.0",
-    "@typescript-eslint/parser": "^4.32.0",
+    "@eslint/js": "^9.0.0",
+    "@types/cli-progress": "^3.11.0",
+    "@types/dockerode": "^4.0.0",
+    "@types/fs-extra": "^11.0.0",
+    "@types/inquirer": "^9.0.0",
+    "@types/jest": "^29.0.0",
+    "@types/lodash": "^4.17.0",
+    "@types/node": "^24.0.0",
+    "@types/npm-registry-fetch": "^8.0.0",
+    "@types/semver": "^7.7.0",
+    "@types/spdx-correct": "^3.1.0",
+    "@types/string-template": "^1.0.0",
+    "@types/tmp": "^0.2.0",
+    "@types/ws": "^8.18.0",
     "copyfiles": "^2.4.1",
-    "eslint": "^7.32.0",
+    "eslint": "^9.0.0",
     "i": "^0.3.7",
-    "jest": "^27.2.4",
+    "jest": "^29.0.0",
     "json": "^11.0.0",
     "performance-now": "^2.1.0",
-    "rimraf": "^3.0.2",
-    "sort-package-json": "^1.52.0",
-    "ts-jest": "^27.0.5",
-    "ts-node": "^10.2.1",
-    "typescript": "^4.4.3"
+    "rimraf": "^6.0.0",
+    "sort-package-json": "^3.0.0",
+    "ts-jest": "^29.0.0",
+    "ts-node": "^10.9.0",
+    "typescript": "^5.0.0",
+    "typescript-eslint": "^8.0.0"
   },
   "dxw": {
     "commands": [
