@@ -5,8 +5,8 @@ describe('test analyse for default plugins', () => {
 
     runOnlyLocally('test analyse for javascript and ruby', async () => {
 
-        await analyseFiles(['/Users/mario/test-projects/depinder/dxworks'], {results:'results-test-mongo', refresh: false,
-            plugins: ['.net']})
+      await analyseFiles(['/Users/mario/test-projects/depinder/dxworks'], {results:'results-test-mongo', refresh: false,
+        plugins: ['.net']})
 
         console.log('done')
     }, 7200000)
