@@ -1,11 +1,11 @@
 import {dotnet, NugetRegistrar} from '../src/plugins/dotnet'
-import minimatch from 'minimatch'
+import { minimatch } from 'minimatch'
 
 describe('default test', () => {
     it('should pass', async () => {
         const res = await new NugetRegistrar().retrieve('Unity')
 
-        console.log(res)
+        // console.log(res)
     })
 
     it('should match just files with *proj extension', async () => {
